backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "keur-doctor-backend"
version = "0.1.0"
description = "FastAPI backend for Keur Doctor scheduling SaaS"
authors = [{ name = "Keur Doctor" }]
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.30.0",
  "SQLAlchemy[asyncio]>=2.0.28",
  "asyncpg>=0.29.0",
  "alembic>=1.13.1",
  "psycopg[binary]>=3.1.18",
  "httpx>=0.27.0",
  "cachetools>=5.3.2",
  "pyjwt[crypto]>=2.8.0",
  "casbin>=1.35.0",
  "casbin-sqlalchemy-adapter>=1.3.0",
  "pydantic-settings>=2.2.1",
  "slowapi>=0.1.9",
  "structlog>=24.1.0",
  "prometheus-client>=0.20.0",
  "opentelemetry-sdk>=1.24.0",
  "opentelemetry-exporter-otlp>=1.24.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.1",
  "pytest-asyncio>=0.23.5",
  "ruff>=0.3.0",
  "mypy>=1.9.0",
  "types-cachetools>=5.3.0.8",
]

[project.scripts]
keur-doctor-backend = "src.main:run"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
strict = false

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src", "README.md", "casbin", "alembic"]
